Update Python dependencies
Django 3.0 has come out so this change limits updates to 2.2.* because
it is too close to the show to be updating to the next major version.
The motivation for doing this update at all is that django-ajax-selects
needed to be updated to 1.8.0 to fix compatibility with the earlier
Django 2.2 update.